+++ client/src/components/UserTrackContainer/UserTrackContainer.css
@@ -8,3 +8,7 @@
   display: flex;
   flex-direction: column;
 }
+
+.user-track-container .loading {
+  margin-top: 100px;
+}
